Move settings page to features and add lazy load
